Use `development` for the webpack mode even in production
With devtools set to `source-map` and mode to `production` it's
impossible as of April 2024 to add breakpoints in both firefox and
chrome.

It does lead to a somewhat small file but at the cost of debugability.
When using `eval-source-map` for devtools no matter what is the mode you
can properly debug but the resulting file is fairly large (800K).

Finally using mode `development` allows to debug and the resulting file
is "only" twice as large so we will settle on that.

Signed-off-by: Matthieu Patou <mat@matws.net>