Improve date handling
Previous to this change the plugin assumed that the date should be
displayed in the timezone of the browser no matter what was the setting.

Also it wouldn't deal with the timerange that was specified (sic).

This is fixed in this change now, we leverage `serverDelay` to send the
metrics back in time by providing a positive delay and forward in time
(in case the selected timezone is ahead of time) by providing a negative
delay.