Avoid continuous redraw of the component
With recent versions of Grafana (11.5+ at least) I noticed that changing
the title of the graph would cause the whole component to be redrawn
which is kind of costly especially if you have a lot of horizons.

Switched to a debounced rendering pattern where we don't redraw right
away most of the time instead we fire a timer with a callback to
actually do the rendering.

Once in a while (ie. 1sec) we force rerendering to ensure that things
are still kept up to date.

Signed-off-by: Matthieu Patou <mat@matws.net>